backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
sentence-transformers
faiss-cpu
rank-bm25
python-dotenv
